超时跳转登录，根据用户信息确认地图中心点

+++ service-map/src/main/java/com/yuyu/map/dao/GisMapper.java
@@ -10,5 +10,5 @@ public interface GisMapper {
 
     List<GIS> getAll();
 
-    List<GIS> getByTreeIds(List<Long> treeIds);
+    List<GIS> getByTreeIds(List<Integer> treeIds);
 }
